// apps/pds/src/types/components/attributes.ts
export type MyComponentSize = 'xs' | 'sm' | 'md' | 'lg'
export type MyButtonComponentVariant =
  | 'default'
  | 'primary'
  | 'secondary'
  | 'warning'
  | 'danger'
  | 'light'
  | 'theme'

export type MyTextFieldComponentVariant =
  | 'default'
  | 'primary'
  | 'secondary'
  | 'warning'
  | 'danger'
  | 'light'
  | 'simple'

export type MyComponentFormValues = 'children' | 'educVocational'